Added the MCMC code and the halo profiles
This commit is for adding the MCMC estimation script (to be merged into the main code preferably) along with some files where the code for the halo profiles is stored, which the MCMC code depends on. `Dsigma_redmagic_0.01.dat` is an example catalog file for testing MCMC provided by Surhud, which is also included.