working files for the MCMC est.
For adding the MCMC estimation script (to be merged into the main code preferably) along with some files where the code for the halo profiles is stored, which the MCMC code depends on. Dsigma_redmagic_0.01.dat is an example catalog file for testing MCMC provided by Surhud, which is also included. Currently `MCMC_parameter_est.py` uses MPI by default inside the script (`mpiexec` or a similar command will have to be used to run it), but a small part of the code can be uncommented to run without MPI.

No changes to the existing code in the master branch have been made.